tests/wpt/web-platform-tests/css/css-backgrounds/parsing/background-attachment-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Backgrounds and Borders Module Level 3: parsing background-attachment with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-backgrounds/#background-attachment">
<meta name="assert" content="background-attachment supports the full grammar '<attachment>#'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("background-attachment", "fixed");
test_valid_value("background-attachment", "scroll, fixed, local, fixed, scroll");
</script>
</body>
</html>
